Updated PS2000A _lowLevelGetTriggerTimeOffset
Fixes #85.
segmentIndex uses c_uint32 now instead of c_uint16
As confirmed by picotech, this is in fact an issue with the documentation. They ps2000aApi.h is correct and the pdf has a typo.

+++ picoscope/ps2000a.py
@@ -393,7 +393,7 @@ class PS2000a(_PicoscopeBase):
             byref(timeUpper),
             byref(timeLower),
             byref(timeUnits),
-            c_uint16(segmentIndex),
+            c_uint32(segmentIndex),
             )
         self.checkResult(m)
